broke generator code out into their own files based on purpose.  Also replaced spaces with tabs.  Also disabled the obsolete dragon buy/grind macro.  The old krng logic was coupled with the generator code.  It's also coupled with nox.py via the points.  I'll set on decoupling those two later.
