+++ .gitignore
@@ -1,1 +1,3 @@
 *.pyproj
+__pycache__/
+.vscode/
